requirements.txt
twilio~=9.5.1
python-dotenv~=1.1.0
rich~=14.0.0

Flask~=3.1.0